Car Suspension | Positioning System
Car Suspension
To simulate the behavior of a car suspension system when the car is driven over an obstacle. Different situations, for example when the speed or the loading of the car changes was examined. A simplified model of a car suspension system is shown in 'Car Suspension Overview' file. The quarter of the car body mass (M) is taken into consideration but the half of the axle mass together with the mass of the wheels is neglected. Newton's law in conjunction with laplace transform was applied to obtain a transfer function which was implemented in MATLAB Simulink