Torpedo
Torpedo

Control system used to stabilise the response of the torpedo to a step change in depth setting. Tuned the system parameters to obtain a satisfactory respons.
Simulink model was designed, neglecting the gravity effects.

'Torpedo_3' file was used to trail and approach
'storpedo' file simulates the non-linear system
'torpedo.m' is the main file